perf(ui): isolate session prefix resolution
